ajout et correction d'écran + modification de certains model et controllers

+++ go/internal/models/productModel.go
@@ -5,7 +5,7 @@ type Product struct {
 	Name         string `gorm:"size:64; not null" json:"name"`
 	Picture      string `gorm:"size:100;" json:"picture"`
 	Type         string `gorm:"size:100; not null" json:"type"`
-	JetonsRequis uint   `gorm:"default: 0; not null" json:"jetons_requis"`
-	Nb_Products  uint64 `gorm:"default:0; not null" json:"nb_products"`
-	StandID      uint64 `gorm:"not null" json:"stand_id"` // Clé étrangère vers le stand
+	JetonsRequis int    `gorm:"default: 0; not null" json:"jetons_requis"`
+	Nb_Products  int    `gorm:"default:0; not null" json:"nb_products"`
+	StandID      int    `gorm:"not null" json:"stand_id"` // Clé étrangère vers le stand
 }
